pyproject.toml
[project]
name = "ttfm-eval"
version = "0.1.0"
description = "Standalone evaluation infrastructure for TTFM and forecasting baselines"
readme = "README.md"
requires-python = ">=3.11, <3.14"
dependencies = [
    "chronos-forecasting",
    "numpy",
    "openai>=1.0.0",
    "pandas",
    "torch",
    "tqdm",
    "sentence-transformers>=5.1.1",
    "timesfm[torch]",
    "tabpfn>=6.3.1",
    "tabpfn-time-series>=1.0.7",
    "prophet>=1.2.2",
    "requests>=2.32.5",
    "nest-asyncio>=1.5.0",
]

[tool.uv.sources]
# Pin TimesFM to same git rev as training repo so TimesFM_2p5_200M_torch API is available.
# PyPI timesfm may use a different API and cause "module 'timesfm' has no attribute 'TimesFM_2p5_200M_torch'".
timesfm = { git = "https://github.com/google-research/timesfm.git", rev = "bf88c5dc88f6275b7071ebcd9051406c11435c0c" }

[project.optional-dependencies]
# Toto baseline (install with: uv sync --extra toto). May conflict with torch/vllm versions.
# toto = ["toto-ts"]
# vLLM for TTFM context summarizer and gpt_forecast (install if using those baselines).
vllm = ["vllm>=0.11.0"]

[project.scripts]
ttfm-eval = "ttfmeval.evaluation:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ttfmeval"]

[tool.hatch.build.targets.sdist]
packages = ["src/ttfmeval"]
